requirements.txt


# database

MySQL-python == 1.2.5


# web

web.py == 0.37

# sina app engine
sae-python-dev == 1.3.6

# template
Jinja2 == 2.8

# wechat
WeRoBot == 0.6.0

